+++ api/src/main/java/me/combimagnetron/sunscreen/neo/render/engine/encode/MapEncoder.java
@@ -153,7 +153,7 @@ public class MapEncoder {
                         out.writeBits(4, index & BinaryMasks.FOUR_BIT_MASK);
                     }
                 }
-                out.writeBits(32, Float.floatToIntBits(renderChunk.scale()));
+                out.writeBits(32, Float.floatToIntBits(renderChunk.scale().floatValue()));
                 out.writeBits(32, Float.floatToIntBits(renderChunk.position().x()));
                 out.writeBits(32, Float.floatToIntBits(renderChunk.position().y()));
                 out.writeBits(32, Float.floatToIntBits(renderChunk.position().z()));
